updated order ctrl

+++ controllers/productctrl.js
@@ -9,33 +9,38 @@ const handleProductUpload = async (req, res) => {
   const { name, category, image, price, description, stores } = req.body;
   const { id } = req.params;
   try {
-    //find user with ID and check if user is an admin
-    const isAdmin = await userModel.findById(id);
-    if (isAdmin && isAdmin.role.toLowerCase() === "admin") {
-      const exists = await productModel.findOne({ name: name });
-      if (exists) {
-        res.send({ message: "Product Already Listed" });
-      } else {
-        if (name && category && image && price && description && stores) {
-          const imageUpload = await cloudinary.uploader.upload(image, {
-            folder: "Hcue",
-            timeout: 60000,
-          });
+    if(id){
 
-          const storeList = stores.map((item) => item.value);
-
-          const data = await productModel.create({
-            name,
-            category,
-            image: imageUpload?.secure_url,
-            price,
-            description,
-            stores: storeList,
-          });
-          res.send({ message: "Upload successfully", data });
+      //find user with ID and check if user is an admin
+      const isAdmin = await userModel.findById(id);
+      if (isAdmin && isAdmin.role.toLowerCase() === "admin") {
+        const exists = await productModel.findOne({ name: name });
+        if (exists) {
+          res.send({ message: "Product Already Listed" });
+        } else {
+          if (name && category && image && price && description && stores) {
+            const imageUpload = await cloudinary.uploader.upload(image, {
+              folder: "Hcue",
+              timeout: 60000,
+            });
+  
+            const storeList = stores.map((item) => item.value);
+  
+            const data = await productModel.create({
+              name,
+              category,
+              image: imageUpload?.secure_url,
+              price,
+              description,
+              stores: storeList,
+            });
+            res.send({ message: "Upload successfully", data });
+          }
         }
+      } else {
+        res.send({ message: "only admins can perform this action" });
       }
-    } else {
+    }else{
       res.send({ message: "only admins can perform this action" });
     }
   } catch (error) {
@@ -59,26 +64,31 @@ const handleTurnOffProduct = async (req, res) => {
   const { location, products } = req.body;
   const { id } = req.params;
   try {
-    //find user with ID and check if user is an admin
-    const isAdmin = await userModel.findById(id);
-    if (isAdmin && isAdmin.role.toLowerCase() === "admin") {
-      if (location && products) {
-        products.forEach(async (el) => {
-          const product = await productModel.findById(el.value);
-          const stores = product.stores;
-          console.log(stores);
-          await productModel.updateOne(
-            { _id: product._id },
-            {
-              stores: stores.filter(
-                (el) => el.toLowerCase() !== location.toLowerCase()
-              ),
-            }
-          );
-        });
-        res.send({ message: "availability turned off" });
+    if(id){
+      //find user with ID and check if user is an admin
+      const isAdmin = await userModel.findById(id);
+      if (isAdmin && isAdmin.role.toLowerCase() === "admin") {
+        if (location && products) {
+          products.forEach(async (el) => {
+            const product = await productModel.findById(el.value);
+            const stores = product.stores;
+            console.log(stores);
+            await productModel.updateOne(
+              { _id: product._id },
+              {
+                stores: stores.filter(
+                  (el) => el.toLowerCase() !== location.toLowerCase()
+                ),
+              }
+            );
+          });
+          res.send({ message: "availability turned off" });
+        }
+      } else {
+        res.send({ message: "only admins can perform this action" });
       }
-    } else {
+
+    }else{
       res.send({ message: "only admins can perform this action" });
     }
   } catch (error) {
@@ -91,24 +101,28 @@ const handleTurnOnProduct = async (req, res) => {
   const { id } = req.params;
   console.log(products);
   try {
-    //find user with ID and check if user is an admin
-    const isAdmin = await userModel.findById(id);
-    if (isAdmin && isAdmin.role.toLowerCase() === "admin") {
-      if (location && products) {
-        products.forEach(async (el) => {
-          const product = await productModel.findById(el.value);
-          const stores = product.stores;
-          console.log(stores);
-          if (!stores.includes(location)) {
-            await productModel.updateOne(
-              { _id: product._id },
-              { stores: [...stores, location] }
-            );
-          }
-        });
-        res.send({ message: "availability turned on" });
+    if(id){
+      //find user with ID and check if user is an admin
+      const isAdmin = await userModel.findById(id);
+      if (isAdmin && isAdmin.role.toLowerCase() === "admin") {
+        if (location && products) {
+          products.forEach(async (el) => {
+            const product = await productModel.findById(el.value);
+            const stores = product.stores;
+            console.log(stores);
+            if (!stores.includes(location)) {
+              await productModel.updateOne(
+                { _id: product._id },
+                { stores: [...stores, location] }
+              );
+            }
+          });
+          res.send({ message: "availability turned on" });
+        }
+      } else {
+        res.send({ message: "only admins can perform this action" });
       }
-    } else {
+    }else{
       res.send({ message: "only admins can perform this action" });
     }
   } catch (error) {
@@ -120,19 +134,24 @@ const handleProductDelete = async (req, res) => {
   const { id } = req.params;
   const { body } = req;
   try {
-    //check if user is admin
-    const isAdmin = await userModel.findById(id);
-    if (isAdmin && isAdmin.role.toLowerCase() === "admin") {
-      //loop through data ,find by id and delete
-      if (body) {
-        body.forEach(async (elem) => {
-          await productModel.findByIdAndDelete(elem.value);
-        });
-        //find product Db and send
-        const productdb = await productModel.find();
-        res.send({ data: productdb });
+    if(id){
+      //check if user is admin
+      const isAdmin = await userModel.findById(id);
+      if (isAdmin && isAdmin.role.toLowerCase() === "admin") {
+        //loop through data ,find by id and delete
+        if (body) {
+          body.forEach(async (elem) => {
+            await productModel.findByIdAndDelete(elem.value);
+          });
+          //find product Db and send
+          const productdb = await productModel.find();
+          res.send({ data: productdb });
+        }
+      } else {
+        res.send({ message: "only admins can perform this action" });
       }
-    } else {
+
+    }else{
       res.send({ message: "only admins can perform this action" });
     }
   } catch (error) {
